Moved test case generators to AbstractFunctionTestCase and fixed for aggregations
